scripts/adhoc/requirements_dashboard.txt
# SQL Server Performance Dashboard Requirements
# Install: pip install -r requirements_dashboard.txt

# Core dashboard framework
streamlit>=1.28.0

# Data visualization
plotly>=5.17.0

# Data processing
pandas>=2.1.0
numpy>=1.24.0

# Database connectivity
pyodbc>=5.0.0

# Environment variables
python-dotenv>=1.0.0

# Optional: For better date handling
python-dateutil>=2.8.2
